Move Control.Parallel to TaskPool effect and clean up threads
