+++ hscli.cabal
@@ -93,10 +93,11 @@ library
     App.Scan.Project
     App.Scan.ProjectInference
     Control.Carrier.Output.IO
+    Control.Carrier.TaskPool
     Control.Carrier.Threaded
     Control.Effect.Output
+    Control.Effect.TaskPool
     Control.Effect.Threaded
-    Control.Parallel
     DepTypes
     Discovery
     Discovery.Walk
